Consistency fixes
Fixed consistency getting reread every tic in multi-tick packets. Removed quick ticking if a network tic is available.

+++ src/g_game.cpp
@@ -1129,7 +1129,7 @@ static uint32_t StaticSumSeeds()
 		pr_damagemobj.Seed();
 }
 
-static uint32_t CalculateConsistency(int client, uint32_t seed)
+static int16_t CalculateConsistency(int client, uint32_t seed)
 {
 	if (players[client].mo != nullptr)
 	{
